Update for 24-25
- VS2022 platform toolset v142->v143
- Teacher naming
- Fix build errors (precompiled headers)
- Enable correct C++ version

+++ Week2/HelloReplaceString/HelloReplaceString.cpp
@@ -1,5 +1,9 @@
-#include <iostream>
-#include <string>
+/*#include <iostream>
+#include <string>*/
+
+import std.core;
+
+
 using namespace std;
 
 // Replace oldSubStr in s with newSubStr
@@ -25,12 +29,12 @@ bool replaceString(string& s, const string& oldSubStr, const string& newSubStr) 
 	bool isReplaced = false;
 	int currentPosition = 0;
 	while (currentPosition < s.length()) {
-		int position = s.find(oldSubStr, currentPosition);
+		int position = (int)s.find(oldSubStr, currentPosition);
 		if (position == string::npos) // Mo more matches
 			return isReplaced;
 		
 		s.replace(position, oldSubStr.length(), newSubStr);
-		currentPosition = position + newSubStr.length();
+		currentPosition = position + (int)newSubStr.length();
 		isReplaced = true; // At least one match
 	
 	}
